+++ web/src/lib/components/asset-viewer/asser-viewer-nav-bar.svelte
