+++ web/src/lib/components/asset-viewer/asset-viewer.svelte
