+++ web/src/lib/components/asset-viewer/detail-panel.svelte
