+++ web/src/lib/components/asset-viewer/intersection-observer.svelte
